requirements.txt
flask==2.3.3
google-cloud-datastore==2.15.1
python-dotenv>=0.19.2
authlib>=1.0
six
python-jose
